moved method processing into separated syntax walker
